restore window position on startup, add support for notan 0.10.0

+++ src/utils.rs
@@ -1158,7 +1158,7 @@ impl ImageExt for RgbaImage {
 
     fn to_texture(&self, gfx: &mut Graphics) -> Option<Texture> {
         gfx.create_texture()
-            .from_bytes(self, self.width() as i32, self.height() as i32)
+            .from_bytes(self, self.width(), self.height())
             .with_mipmaps(true)
             .with_format(notan::prelude::TextureFormat::SRgba8)
             // .with_premultiplied_alpha()
@@ -1170,7 +1170,7 @@ impl ImageExt for RgbaImage {
 
     fn to_texture_premult(&self, gfx: &mut Graphics) -> Option<Texture> {
         gfx.create_texture()
-            .from_bytes(self, self.width() as i32, self.height() as i32)
+            .from_bytes(self, self.width(), self.height())
             .with_premultiplied_alpha()
             // .with_filter(TextureFilter::Linear, TextureFilter::Nearest)
             // .with_wrap(TextureWrap::Repeat, TextureWrap::Repeat)
